Extracted interface
Went through code and replaced usages of the solid class with the
interface. In the tests, created a MockDiceSet and allowed it to accept
parameters of the values it should contain. The tests all pass, Dice (or
IDiceSet as I referred to it) was not the the thing being tested, so the
Mock implementation didn't matter.